The ColorScheme name in kdeglobals is not the filename.
We need to scan all color schemes to see if it matches the `[General] Name=`.

I made it first look for a file with the exact same name as an optimization,
but it will still open up every color scheme in the home directory before
moving on to the root directory.

Further optimization could list every color scheme in all 4 directories,
then prioritize looking at files with the same name before moving on to
the other results.