Override Breeze desktoptheme with Papirus icons
We create empty SVGs for most icons so it falls back to the Papirus
icon theme. There are a few unique icons in the desktoptheme that we
need to set though. I've grabbed the necessary icons.

Plasma loads the icons by taking the icon name "audio-volume-high"
and looking for a file removing everything after the first dash "-".
So it looks for "audio.svg", then finds an element in the SVG with
id="audio-volume-high".

After adding the Papirus icon rects/paths, I adjusted the fill to
style="fill:currentColor" and added the class="ColorScheme-Text".

Finally, I noticed that the icons were larger since they were losing
the whitespace around them. I needed to add an invisible rect to set
the icon width/height.